+++ src/components/ui/badge/index.js
@@ -15,6 +15,8 @@ export const badgeVariants = cva(
           "border-transparent bg-destructive text-white [a&]:hover:bg-destructive/90 focus-visible:ring-destructive/20 dark:focus-visible:ring-destructive/40 dark:bg-destructive/60",
         outline:
           "text-foreground [a&]:hover:bg-accent [a&]:hover:text-accent-foreground",
+        "genre-solid":
+          "inline-block text-xs font-sans px-2 py-1 rounded-full border-0 bg-white text-black",
       },
     },
     defaultVariants: {